baselines/BERT-LS/requirements.txt
python-magic==0.4.24 # detect file encoding
pandas==0.25.1
tqdm==4.64.0
nltk==3.4.5
ftfy
boto3
requests
torch
regex
sklearn
openpyxl
